refactor(backendpokemonapi/src/modules); organized scaffolding
