+++ backend-pokemon-api/src/modules/auth/auth.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
-import { TrainerRepository } from 'src/trainers/trainer.repository';
-import { Trainer } from 'src/trainers/entities/trainer.entity';
+import { TrainerRepository } from 'src/modules/trainers/trainer.repository';
+import { Trainer } from 'src/modules/trainers/entities/trainer.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { TrainerService } from 'src/trainers/trainer.service';
-import { TrainerModule } from 'src/trainers/trainer.module';
+import { TrainerService } from 'src/modules/trainers/trainer.service';
+import { TrainerModule } from 'src/modules/trainers/trainer.module';
 import { AuthController } from './auth.controller';
 
 @Module({
