+++ backend-pokemon-api/src/modules/trainers/entities/trainer.entity.ts
