+++ backend-pokemon-api/src/modules/trainers/trainer.module.ts
