+++ backend-pokemon-api/src/modules/trainers/trainer.repository.ts
