+++ backend-pokemon-api/src/modules/trainers/trainer.service.ts
